requirements.txt
# ============================================================
# AI Hiring Agent System - Python Dependencies
# requirements.txt
# ============================================================

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0          # PostgreSQL async driver
alembic==1.13.2          # Migrations

# Settings
pydantic==2.9.0
pydantic-settings==2.5.0

# AI / NLP / LLM
transformers==4.44.0
torch==2.4.0
sentence-transformers==3.1.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
openai==1.44.0

# CV Parsing
pdfminer.six==20240706
docx2txt==0.8
python-docx==1.1.2

# RAG / Vector DB
chromadb==0.5.0
faiss-cpu==1.8.0

# Video Processing
opencv-python-headless==4.10.0.84
deepface==0.0.92
numpy==1.26.4

# Storage
boto3==1.35.0

# Communication
twilio==9.3.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Email monitoring (for email CV submissions)
imaplib2==3.6

# Background tasks
celery==5.4.0
redis==5.0.8

# Dev / Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
